@@ -1,3 +1,3 @@
 module github.com/zituocn/logx
 
-go 1.16
+go 1.20
